Update LaunchConfig and how we install Tor
Switch to using pipenv to install python packages on the instance and
run the monitor as a python module.

Tor is now installed and started as a background service using the
steps recommended by the Tor Project ubuntu guide. This also adds a
check to the UserData that Tor is running (on 9050) before proceeding
with the instance launch.